Better handling of type annotations. These changes follow on recent changes to AFU.
Type annotations appear in the ATypeElement that is in the .type slot of an
AElement -- not directly in the AElement.
Update test goals to reflect that type annotations go in type subfield.

Use LinkedHashMap and LinkedHashSet in an attempt to make output more deterministic.

Add more tests.

TypePrinter.visitClassType handles baseType in Java or JVML format

Improve debugging output.